Created parsing/test_util
Moved tests specific to parsing utilities to their own test module.
